requirements.txt

# Base ----------------------------------------
numpy
opencv-python>=4.1.1
Pillow>=7.1.2
PyYAML>=5.3.1
requests>=2.23.0
scipy>=1.4.1
torch>=1.7.0,!=1.12.0
torchvision>=0.8.1,!=0.13.0
tqdm>=4.41.0
protobuf<4.21.3
fastapi
uvicorn 
python-multipart 
pyjwt>=2.10.1
bcrypt>=4.2.1
passlib>=1.7.4
azureml
azureml-core>=1.32.0
# Logging -------------------------------------
tensorboard>=2.4.1
# wandb


# AZURE ------------------------------------
azure-keyvault>=4.2.0
azure-identity



pandas>=1.1.4
seaborn>=0.11.0
python-dotenv


ipython 
psutil  
thop  

